Create README.md
This code controls the colors of a RGB LED stripe using an ATMEGA 644P and a TIP31.
The colors are sent using the UART port. 
electronics will be added later
Example:
B005 sets the pwm of the blue color to 5. 
B040 sets the pwm to 40. 
R255 sets the pwm of the red color to 255 which is the max value.
g000 sets the pwm of the green color to 0 which means off.

There are more commands.